Created Interact system
Created input action for interacting, and created the actor system for being interacted with as well as the player line tracing to interact. Also created park scene and created basic UE landscape.

Implemented Resource map in the game instance and made the main menu. Main menu still needs programming inside it for when we have the intro to the game completed.